refactor utils to seperate file
