+++ src/MoneyInput/MoneyInput.tsx
@@ -1,5 +1,6 @@
 import React, { useState, useEffect, useMemo } from 'react'
 import _styles from './MoneyInput.module.css'
+import { convertToCents, sanitizeValue } from './utils'
 
 interface MoneyInputProps extends React.InputHTMLAttributes<HTMLInputElement> {
   locale?: 'de-DE' | 'en-US' | 'en-GB' | 'fr-FR'
@@ -35,21 +36,6 @@ export default function MoneyInput({
       }
     }
   }, [value, currencyFormatter, isFocused])
-
-  const convertToCents = (value: number): number => {
-    const centsValue = Math.round(value * 100)
-    return centsValue
-  }
-
-  const sanitizeValue = (value: string) => {
-    value = value.replace(/[^\d,.]+/g, '')
-    const parts = value.split(/[,.]/)
-    if (parts.length > 2) {
-      value = parts.slice(0, -1).join('') + '.' + parts.slice(-1)
-    }
-    const sanitizedValue = parseFloat(value.replace(/,/g, '.'))
-    return !isNaN(sanitizedValue) ? sanitizedValue : 0
-  }
 
   const handleFocus = () => {
     setInputValue((value / 100).toString())
